moved functions and variables for customizing tier list to a js file
