+++ Tier-list-website/Tier-List-Clone-Website/src/views/CustomizeScreen.vue
@@ -3,116 +3,8 @@ import draggable from 'vuedraggable'
 import TierListDisplay from '../components/TierListDisplay.vue'
 import ModifyTierList from '../components/ModifyTierList.vue'
 import DeleteTiers from '../components/DeleteTiers.vue'
-import {ref, onMounted} from 'vue'
-
-// Contains the initial default tier list
-var default_tier_list = ref([
-    {
-        tier_name: "S", 
-        color: "rgb(253, 123, 123)",
-        tier_image_container: []
-    },
-    {
-        tier_name: "A", 
-        color: "rgb(255, 170, 91)",
-        tier_image_container: []
-    },
-    {
-        tier_name: "B", 
-        color: "rgb(253, 226, 144)",
-        tier_image_container: []
-    },
-    {
-        tier_name: "C", 
-        color: "rgb(246, 253, 123)",
-        tier_image_container: []
-    },
-    {
-        tier_name: "D", 
-        color: "rgb(146, 246, 115)",
-        tier_image_container: []
-    },
-    {
-        tier_name: "F", 
-        color: "rgb(123, 253, 244)",
-        tier_image_container: []
-    },
-])
-
-// Contains colors used in default tier list
-const colors_of_default_tier_list = ["rgb(253, 123, 123)", "rgb(255, 170, 91)", "rgb(253, 226, 144)", "rgb(246, 253, 123)", "rgb(146, 246, 115)", "rgb(123, 253, 244)"]
-
-// Opens modal dialog for customizing tier list
-var open_modal_dialog = ref(false)
-
-// Opens modal dialog for deleting tiers in tier list
-var delete_tiers_modal_dialog = ref(false)
-
-// Contains image files
-var files_arr = ref([])
-
-// Inserts a new tier in the tier list
-function add_new_tier(tier_list){
-    const num_of_tiers = tier_list.length
-    tier_list.push({
-        tier_name: `Tier ${num_of_tiers+1}`,
-        color: colors_of_default_tier_list[getRandomInt(0, colors_of_default_tier_list.length)],
-        tier_image_container: []
-    })
-}
-
-// Gets a random number between the minimum number and maximum number
-function getRandomInt(min, max) {
-  const minCeiled = Math.ceil(min)
-  const maxFloored = Math.floor(max)
-  return Math.floor(Math.random() * (maxFloored - minCeiled) + minCeiled)
-}
-
-// Updates current tier list based on modifications
-function updateTierList(state, new_tier_list){
-    open_modal_dialog.value = state
-    default_tier_list.value = new_tier_list
-}
-
-// Deletes selected tiers from current tier list
-function deleteTiers(state, deleted_tiers){
-    delete_tiers_modal_dialog.value = state
-    const new_tier_list = default_tier_list.value.filter((tier) => !deleted_tiers.includes(tier))
-    default_tier_list.value = new_tier_list
-}
-
-// Upload images to image container
-function uploadToImageContainer(){
-    // Creates hidden input element that allows the user to upload their images
-    const access_files = document.createElement("input");
-    access_files.type = "file";
-    access_files.accept = "image/png, image/jpeg";
-    access_files.multiple = true;
-    access_files.style.visibility = 'hidden'; 
-
-    // Executes when user selects their images
-    access_files.onchange = () => {
-        // Array of selected files
-        const files = access_files.files;
-
-        // Accesses the div where all images will be uploaded to
-        const placeHolder = document.getElementById('place-holder');
-
-        // Iterate through every file
-        for(const file of files){
-            // Reads current file
-            const reader = new FileReader();
-            reader.onload = (e) => {
-                // Append image file to an array of draggable elements
-                files_arr.value.push(e.target.result)
-            }
-            // Reads contents of current file
-            reader.readAsDataURL(file);
-        } 
-    }
-    // Simulate mouse click
-    access_files.click();
-}
+import {default_tier_list, open_modal_dialog, delete_tiers_modal_dialog, files_arr} from '../front-end-code/customize_screen_functions'
+import {add_new_tier, updateTierList, deleteTiers, uploadToImageContainer} from '../front-end-code/customize_screen_functions'
 </script>
 
 <template>
@@ -145,6 +37,7 @@ function uploadToImageContainer(){
                 </div>
                 <div class="d-flex text-start justify-start align-start h-100 w-100" id="place-holder">
                     <draggable
+                    group="tier_list"
                     :list="files_arr"
                     itemKey="index">
                         <template #item="{ element }">
